requirements.txt
opencv-python
numpy==1.14.3
scipy==1.1.0
Pillow # For JPEG model
tensorflow-gpu==1.13.1
torch==0.4.1 # For Denoiser model
